+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - The admin web UI now shows a list of all tenants which have been registered to the site under
   `/admin/tenant`. So far Forge Cloud does not submit tenant information. Therefore the list will be empty.
 - The admin web UI now shows a list of all tenant keysets under `/admin/tenant_keyset`.
+- The site controller has a new "VPC prefix" resource type, which is analogous
+  to a network segment in L3 VPCs.
 
 ### Changed